refactor: move module finalization into apply_transfer_parts_mutation
